add utils.py with log_progress and line counts
